perf: dynamic scheduling, Node-struct shrink, and a redundant-parse fix
Bundles the high-risk trio from feature-candidates.md Section F, scoped
down from their original framing after measurement showed the initial
premise for two of the three didn't hold:

- F2 (full): replaces static contiguous file-chunking in analyze_files
  with a shared-atomic-cursor scheduler walking a size-descending (LPT)
  claim order, so a workspace with one huge file and many tiny ones no
  longer leaves most workers idle. The input-order output guarantee is
  now explicit reassembly by original index rather than an accident of
  contiguous partitioning; mutation-tested against an append-on-arrival
  collector. 4.80x median speedup on a size-skewed workspace, 1.004x
  (no regression) on an even one, byte-identical reports either way.

- F9 (revised): a memory-peak benchmark found parsing cost is strictly
  linear with file size (~30x memory, no wall at realistic sizes) --
  not the superlinear growth that would justify a streaming parser, and
  streaming would cost the whole codebase's random-access tree API to
  fix a constant factor. Shipped a lower-risk alternative instead: the
  Node struct shrunk from 152 to 72 bytes (u32-backed offsets behind
  the existing usize API, boxed reader-prefix data, narrowed
  symbol_offset), roughly halving memory (~30x -> ~15x per source
  byte) with zero public API change and bit-identical parse trees.

- F3 (revised): investigation found the multi-target edit loop's actual
  bottleneck wasn't a single-form incremental-parse opportunity (the
  premise doesn't hold for commands that change a document's top-level
  form count, like splice/split/join/slurp/barf) but a genuinely
  redundant second full parse of the same bytes on every iteration.
  Removed the duplicate parse instead of building incremental parsing:
  every tree is still a full fresh parse, only how often the loop
  parses changed. ~1.9x measured speedup, verified by a 96-case
  differential property test comparing the old and new loops after
  every pass, plus mutation-testing that the tests actually catch a
  reintroduced staleness bug.

All three are scoped conservatively relative to the original candidate
framing once evidence contradicted the premise; see the PR description
for the full reasoning trail.

+++ packages/core/syntax/src/sexpr/reindent.rs
@@ -182,7 +182,7 @@ impl SyntaxTree {
             self.node(child)
                 .span
                 .slice(self.source())
-                .get(self.node(child).symbol_offset..)
+                .get(self.node(child).symbol_offset as usize..)
         });
 
         if let Some(distinguished) = head_text.and_then(body_form_distinguished) {
